% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-georgia.R
\docType{data}
\name{font_georgia}
\alias{font_georgia}
\title{Georgia font name R variable aliases}
\format{length 1 character vector}
\usage{
font_georgia
}
\description{
\code{font_os} == "\code{Georgia}"
}
\keyword{datasets}
