% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-install.R
\name{install_fonts_linux}
\alias{install_fonts_linux}
\title{Import fonts in the rlheme package system-wide (on Linux)}
\usage{
install_fonts_linux()
}
\description{
Import fonts in the rlheme package system-wide (on Linux)
}
