% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-lato.R
\docType{data}
\name{font_lato}
\alias{font_lato}
\title{Lato font name R variable aliases}
\format{length 1 character vector}
\usage{
font_lato
}
\description{
\code{font_lato} == "\code{Lato}"
}
\keyword{datasets}
