% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_rl}
\alias{palette_rl}
\title{Redpill-Linpro color palette with 5 qualitative colors}
\usage{
palette_rl(n = 7, name = "RL", type = c("qual", "seq", "div"))
}
\arguments{
\item{n}{number of colors to use (1..5), default is to return all five}

\item{name}{name of palette}

\item{type}{nature of data, default "qual" for qualitative or "seq" for
sequential or "div" for diverging}
}
\value{
named vector with colors and hex codes
}
\description{
This function aligns its signature with RColorBrewer::color.pal()
}
\examples{
palette_rl(1)  # return the primary color
library(scales)
alpha(palette_rl(), 0.4)  # return light signature colors (40 \% alpha)
show_col(alpha(palette_rl(), 0.8))  # return medium light 80\% alpha palette
}
