% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_rl_info}
\alias{palette_rl_info}
\title{Redpill-Linpro color palette information}
\usage{
palette_rl_info()
}
\value{
data frame with information about the available palette(s)
}
\description{
This function aligns its signature with RColorBrewer::color.pal.info()
}
