% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{rl_pal}
\alias{rl_pal}
\title{A qualitative color palette with colors from Redpill-Linpro's graphical profile}
\usage{
rl_pal()
}
\description{
A qualitative color palette with colors from Redpill-Linpro's graphical profile
}
\examples{
library(scales)
scales::show_col(rl_pal()(5))
}
