% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rltheme-package.R
\name{rltheme-exports}
\alias{rltheme-exports}
\alias{\%>\%}
\title{rltheme exported operators}
\description{
The following functions are imported and then re-exported
from the rltheme package to enable use of the magrittr
pipe operator with no additional library calls
}
