% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rltheme-package.R
\docType{package}
\name{rltheme}
\alias{rltheme}
\title{Theme and Theme Components for 'ggplot2' for Redpill-Linpro's graphical profile}
\description{
The core theme: \code{theme_rl} uses Lato
}
\details{
There is an option \code{rltheme.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\keyword{internal}
